Add dbus-glib-1 to telepathy-glib pkg-config check
In recent linker versions, transitive imports of library symbols are
rejected by the linker. dbus-glib-1 exists in the Requires.private
variable of the telepathy-glib pkg-config file, so results in a
transitive import of dbus-glib-1, which fails during the link phase as
dbus_g_type_specialized_construct is called from
server/vino-tube-server.c. Underlinking is another term for this
problem.

Explicitly adding dbus-glib-1 to the telepathy-glib pkg-config check
fixes the problem, bug 646018. Also see
http://blog.flameeyes.eu/2010/11/26/it-s-not-all-gold-that-shines-why-underlinking-is-a-bad-thing