Reset sockets when changing ports, bug 668187
Explicitly remove the IO watches on the server sockets before setting up
new watches, to handle the case where the file descriptor for the socket
has changed. This can occur if during initialisation the bind for
either the IPv4 or IPv6 socket fails, and then the port is changed while
the server is running to a port where both the IPv4 and IPv6 bind
succeeds. In this case, the address family with the failed bind will
have an invalid file descriptor and the socket would not have an
attached IO watch.