libvncserver: fix non-literal format string issue
vino is built with -Werror=format-nonliteral by default and uses a
non-literal format string.

Tag the function that it is used from so that the compiler allows it.
This has a side effect of ensuring that the users of this function are
doing it correctly.

The easiest way to do this correctly is to use the macro in glib, so we
must include the header from this file (which did not yet have it).

https://bugzilla.gnome.org/show_bug.cgi?id=726002